backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.10.0
pydantic-settings>=2.5.0
supabase>=2.7.2
groq>=0.11.0
google-cloud-texttospeech>=2.17.2
redis>=5.1.0
textblob>=0.18.0
python-multipart>=0.0.9
httpx>=0.27.2
python-dotenv>=1.0.1
ruff>=0.6.4
pytest>=8.3.3
pytest-asyncio>=0.24.0
